Phase 3: 新增 user-service.ts（getOrCreateUser/maybeResetBudget/updateUserPlan），key-service 瘦身为 createKey/revokeKey/updateKeyName，critical-write-paths 改为 updateUserBudgetWithAuditTx + insertRequestUsageAndChargeTx(userId)，PG/MySQL 条件 UPDATE 防并发
