+++ packages/core/src/db/d1/critical-writes.impl.ts
@@ -2,10 +2,9 @@
  * D1：关键写路径（batch / 原始 SQL），供 `storage/critical-write-paths` 调度。
  */
 import type { D1PreparedStatement } from '@cloudflare/workers-types';
-import { and, eq } from 'drizzle-orm';
+import { eq } from 'drizzle-orm';
 import type { InsertApiKeyBudgetAuditLogParams } from '../api-key-budget-audit-logs-types';
 import { buildInsertUserAuditLogStatement } from './user-audit-logs.impl';
-import type { InsertUserAuditLogParams } from '../user-audit-logs-types';
 import { insertParamsFromBudgetTx, insertParamsFromCreateKeyAudit, insertParamsFromUsageCharge } from '../user-audit-legacy-mapper';
 import { buildInsertApiKeyStatement } from './api-keys.impl';
 import type { InsertKeyParams } from '../api-keys-types';
@@ -15,7 +14,6 @@ import { roundGatewayMoney } from '../../lib/money-precision';
 import type { D1DatabaseClient } from '../../storage/database-client';
 import { parseMoney } from '../../storage/critical-write-paths-utils';
 import {
-	apiKeysTable as d1ApiKeysTable,
 	systemConfigTable as d1SystemConfigTable,
 	usersTable as d1UsersTable,
 } from '../../storage/drizzle/schema.d1';
@@ -24,24 +22,9 @@ function ensureD1Batch(client: D1DatabaseClient, statements: D1PreparedStatement
 	return client.raw.batch(statements).then(() => undefined);
 }
 
-export async function getActiveApiKeyByUserIdD1(
+export async function getUserBudgetSnapshotD1(
 	client: D1DatabaseClient,
 	userId: string
-): Promise<{ id: string; key: string } | null> {
-	const row = await client.drizzle
-		.select({
-			id: d1ApiKeysTable.id,
-			key: d1ApiKeysTable.key,
-		})
-		.from(d1ApiKeysTable)
-		.where(and(eq(d1ApiKeysTable.userId, userId), eq(d1ApiKeysTable.status, 'active')))
-		.limit(1);
-	return row[0] ?? null;
-}
-
-export async function getApiKeyBudgetSnapshotD1(
-	client: D1DatabaseClient,
-	keyId: string
 ): Promise<{ budgetSpent: number; budgetMax: number | null; budgetPeriod: string | null; budgetResetAt: string | null } | null> {
 	const row = await client.drizzle
 		.select({
@@ -50,9 +33,8 @@ export async function getApiKeyBudgetSnapshotD1(
 			budgetPeriod: d1UsersTable.budgetPeriod,
 			budgetResetAt: d1UsersTable.budgetResetAt,
 		})
-		.from(d1ApiKeysTable)
-		.innerJoin(d1UsersTable, eq(d1ApiKeysTable.userId, d1UsersTable.id))
-		.where(eq(d1ApiKeysTable.id, keyId))
+		.from(d1UsersTable)
+		.where(eq(d1UsersTable.id, userId))
 		.limit(1);
 	if (!row[0]) return null;
 	return {
@@ -86,37 +68,45 @@ export async function createApiKeyWithAuditD1(
 	]);
 }
 
-export async function updateApiKeyBudgetWithAuditTxD1(
+export async function updateUserBudgetWithAuditTxD1(
 	client: D1DatabaseClient,
 	params: {
-		keyId: string;
+		userId: string;
+		/** 读库时的 `users.budget_reset_at`，与之一致才更新（防并发 lazy reset 重复审计） */
+		expectedBudgetResetAt: string | null;
 		budgetSpent: number;
 		budgetResetAt: string | null;
 		budgetMax?: number | null;
+		apiKeyId: string | null;
 		audit: Omit<InsertApiKeyBudgetAuditLogParams, 'id' | 'apiKeyId' | 'afterSpent' | 'afterBudgetResetAt'>;
 	}
 ): Promise<void> {
 	const nextSpent = roundGatewayMoney(params.budgetSpent);
-	const userRow = await client.raw
-		.prepare('SELECT user_id FROM api_keys WHERE id = ?')
-		.bind(params.keyId)
-		.first<{ user_id: string }>();
-	if (!userRow) {
-		throw new Error('updateApiKeyBudgetWithAuditTxD1: api key not found');
-	}
-	const userId = userRow.user_id;
-	const auditRow = insertParamsFromBudgetTx(userId, params.keyId, nextSpent, params.budgetResetAt, params.audit);
+	const auditRow = insertParamsFromBudgetTx(params.userId, params.apiKeyId, nextSpent, params.budgetResetAt, params.audit);
 	const updateStmt =
 		params.budgetMax !== undefined
 			? client.raw
 					.prepare(
-						`UPDATE users SET budget_spent = ?, budget_reset_at = ?, budget_max = COALESCE(?, budget_max), updated_at = datetime('now') WHERE id = ?`
+						`UPDATE users SET budget_spent = ?, budget_reset_at = ?, budget_max = COALESCE(?, budget_max), updated_at = datetime('now') WHERE id = ? AND budget_reset_at IS NOT DISTINCT FROM ?`
 					)
-					.bind(nextSpent, params.budgetResetAt, params.budgetMax == null ? null : roundGatewayMoney(params.budgetMax), userId)
+					.bind(
+						nextSpent,
+						params.budgetResetAt,
+						params.budgetMax == null ? null : roundGatewayMoney(params.budgetMax),
+						params.userId,
+						params.expectedBudgetResetAt
+					)
 			: client.raw
-					.prepare(`UPDATE users SET budget_spent = ?, budget_reset_at = ?, updated_at = datetime('now') WHERE id = ?`)
-					.bind(nextSpent, params.budgetResetAt, userId);
-	await ensureD1Batch(client, [updateStmt, buildInsertUserAuditLogStatement(client.raw, auditRow)]);
+					.prepare(
+						`UPDATE users SET budget_spent = ?, budget_reset_at = ?, updated_at = datetime('now') WHERE id = ? AND budget_reset_at IS NOT DISTINCT FROM ?`
+					)
+					.bind(nextSpent, params.budgetResetAt, params.userId, params.expectedBudgetResetAt);
+	const runResult = await updateStmt.run();
+	const changes = runResult.meta?.changes ?? 0;
+	if (changes === 0) {
+		return;
+	}
+	await ensureD1Batch(client, [buildInsertUserAuditLogStatement(client.raw, auditRow)]);
 }
 
 export async function insertRequestUsageAndChargeTxD1(
